Renamed ScaleManager -> SafeScaleManager Readded older ScaleManager code.
