+++ src/scale-manager/ScaleManager.spec.js
@@ -1,24 +1,15 @@
 import { ScaleManager } from './ScaleManager';
 import { newEvent } from '../debug';
-import { Anchor } from './Anchor';
 
 describe('Scale Manager', () => {
-  const sm = new ScaleManager({ width: 400, height: 400 });
+  const sm = new ScaleManager();
 
   it('Should call the callback passed via the constructor', done => {
-    new ScaleManager({
-      width: 400,
-      height: 400,
-      safeWidth: 200,
-      safeHeight: 200,
-      callback: e => {
-        expect(e.width).to.be.a('number');
-        expect(e.height).to.be.a('number');
-        expect(e.scale).to.be.an('object');
-        expect(e.scale.x).to.be.a('number');
-        expect(e.scale.y).to.be.a('number');
-        done();
-      }
+    const defaultScaleManager = new ScaleManager(e => {
+      expect(e.width).to.be.a('number');
+      expect(e.height).to.be.a('number');
+      expect(e.ratio).to.be.an('number');
+      done();
     });
 
     window.dispatchEvent(newEvent('resize'));
@@ -29,9 +20,7 @@ describe('Scale Manager', () => {
     sm.enable(e => {
       expect(e.width).to.be.a('number');
       expect(e.height).to.be.a('number');
-      expect(e.scale).to.be.an('object');
-      expect(e.scale.x).to.be.a('number');
-      expect(e.scale.y).to.be.a('number');
+      expect(e.ratio).to.be.an('number');
       done();
     });
 
@@ -52,62 +41,5 @@ describe('Scale Manager', () => {
     setTimeout(() => {
       done();
     }, 1);
-  });
-
-  describe('--Anchor Handling--', () => {
-    beforeEach(() => {
-      sm.entities = [];
-    });
-
-    it('should add an anchor', () => {
-      sm.addEntity(new Anchor());
-
-      expect(sm.entities.length).to.equal(1);
-    });
-
-    it('should not add an anchor if it already exists', () => {
-      const testAnchor = new Anchor();
-
-      sm.addEntity(testAnchor);
-      sm.addEntity(testAnchor);
-
-      expect(sm.entities.length).to.equal(1);
-    });
-
-    it('should remove anchors', () => {
-      const testAnchor = new Anchor();
-
-      sm.addEntity(testAnchor);
-      sm.removeEntity(testAnchor);
-
-      expect(sm.entities.length).to.equal(0);
-    });
-
-    it('should invoke anchor update when added', done => {
-      const testAnchor = new Anchor({
-        callback: () => {
-          done();
-        }
-      });
-      sm.addEntity(testAnchor);
-    });
-
-    it('should invoke anchor update on resize', done => {
-      let count = 0;
-
-      const testAnchor = new Anchor({
-        callback: () => {
-          if (count == 1) {
-            done();
-          }
-          count++;
-        }
-      });
-
-      sm.addEntity(testAnchor);
-
-      window.dispatchEvent(newEvent('resize'));
-      // window.dispatchEvent(newEvent('resize'));
-    });
   });
 });
