+++ src/scale-manager/index.js
@@ -1,2 +1,3 @@
 export * from './Anchor';
+export * from './SafeScaleManager';
 export * from './ScaleManager';
